added error handling for handleWebSocketMessage
fixed getBaseUrl
