--- src/features/moshi/components/TextDisplay.tsx
@@ -1,37 +1,0 @@
-import { FC, useEffect, useRef } from "react";
-import { useServerText } from "@/features/moshi/hooks/useServerText";
-
-type TextDisplayProps = {
-  containerRef: React.RefObject<HTMLDivElement>;
-};
-
-export const TextDisplay:FC<TextDisplayProps> = ({
-  containerRef,
-}) => {
-  const { text } = useServerText();
-  const currentIndex = text.length - 1;
-  const prevScrollTop = useRef(0);
-
-  useEffect(() => {
-    if (containerRef.current) {
-      prevScrollTop.current = containerRef.current.scrollTop;
-      containerRef.current.scroll({
-        top: containerRef.current.scrollHeight,
-        behavior: "smooth",
-      });
-    }
-  }, [text]);
-
-  return (
-    <div className="h-full w-full max-w-full max-h-full  p-2 text-white">
-        {text.map((t, i) => (
-          <span
-            key={i}
-            className={`${i === currentIndex ? "font-bold" : "font-normal"}`}
-          >
-            {t}
-          </span>
-        ))}
-    </div>
-  );
-};
