--- src/features/moshi/hooks/useServerAudio.ts
@@ -1,161 +1,0 @@
-import { useCallback, useEffect, useRef, useState } from "react";
-import { useSocketContext } from "@/features/moshi/hooks/SocketContext";
-import { decodeMessage } from "@/features/moshi/protocol/encoder";
-import { useMediaContext } from "@/features/moshi/hooks/MediaContext";
-import { DecoderWorker } from "@/workers/decoderWorker";
-
-export type AudioStats = {
-  playedAudioDuration: number;
-  missedAudioDuration: number;
-  totalAudioMessages: number;
-  delay: number;
-  minPlaybackDelay: number;
-  maxPlaybackDelay: number;
-};
-
-type useServerAudioArgs = {
-  setGetAudioStats?: (getAudioStats: () => AudioStats) => void;
-};
-
-type WorkletStats = {
-  totalAudioPlayed: number;
-  actualAudioPlayed: number;
-  delay: number;
-  minDelay: number;
-  maxDelay: number;
-};
-
-export const useServerAudio = ({setGetAudioStats}: useServerAudioArgs) => {
-  const { socket  } = useSocketContext();
-  const {startRecording, stopRecording, audioContext, worklet, micDuration, actualAudioPlayed } =
-    useMediaContext();
-  const analyser = useRef(audioContext.current.createAnalyser());
-  worklet.current.connect(analyser.current);
-  const startTime = useRef<number | null>(null);
-  const decoderWorker = useRef(DecoderWorker);
-  const [hasCriticalDelay, setHasCriticalDelay] = useState(false);
-  const totalAudioMessages = useRef(0);
-  const receivedDuration = useRef(0);
-  const workletStats = useRef<WorkletStats>({
-    totalAudioPlayed: 0,
-    actualAudioPlayed: 0,
-    delay: 0,
-    minDelay: 0,
-    maxDelay: 0,});
-
-  const onDecode = useCallback(
-    async (data: Float32Array) => {
-      receivedDuration.current += data.length / audioContext.current.sampleRate;
-      worklet.current.port.postMessage({frame: data, type: "audio", micDuration: micDuration.current});
-    },
-    [],
-  );
-
-  const onWorkletMessage = useCallback(
-    (event: MessageEvent<WorkletStats>) => {
-      workletStats.current = event.data;
-      actualAudioPlayed.current = workletStats.current.actualAudioPlayed;
-    },
-    [],
-  );
-  worklet.current.port.onmessage = onWorkletMessage;
-
-  const getAudioStats = useCallback(() => {
-    return {
-      playedAudioDuration: workletStats.current.actualAudioPlayed,
-      delay: workletStats.current.delay,
-      minPlaybackDelay: workletStats.current.minDelay,
-      maxPlaybackDelay: workletStats.current.maxDelay,
-      missedAudioDuration: workletStats.current.totalAudioPlayed - workletStats.current.actualAudioPlayed,
-      totalAudioMessages: totalAudioMessages.current,
-    };
-  }, []);
-
-  const onWorkerMessage = useCallback(
-    (e: MessageEvent<any>) => {
-      if (!e.data) {
-        return;
-      }
-      onDecode(e.data[0]);
-    },
-    [onDecode],
-  );
-
-  let midx = 0;
-  const decodeAudio = useCallback((data: Uint8Array) => {
-    if (midx < 5) {
-      console.log(Date.now() % 1000, "Got NETWORK message", micDuration.current - workletStats.current.actualAudioPlayed, midx++);
-    }
-    decoderWorker.current.postMessage(
-      {
-        command: "decode",
-        pages: data,
-      },
-      [data.buffer],
-    );
-  }, []);
-
-  const onSocketMessage = useCallback(
-    (e: MessageEvent) => {
-      const dataArray = new Uint8Array(e.data);
-      const message = decodeMessage(dataArray);
-      if (message.type === "audio") {
-        decodeAudio(message.data);
-        //For stats purposes for now
-        totalAudioMessages.current++;
-      }
-    },
-    [decodeAudio],
-  );
-
-  useEffect(() => {
-    const currentSocket = socket;
-    if (!currentSocket) {
-      return;
-    }
-    worklet.current.port.postMessage({type: "reset"});
-    console.log(Date.now() % 1000, "Should start in a bit");
-    startRecording();
-    currentSocket.addEventListener("message", onSocketMessage);
-    totalAudioMessages.current = 0;
-    return () => {
-      console.log("Stop recording called in unknown function.")
-      stopRecording();
-      startTime.current = null;
-      currentSocket.removeEventListener("message", onSocketMessage);
-    };
-  }, [socket]);
-
-  useEffect(() => {
-    if (setGetAudioStats) {
-      console.log("Setting getAudioStats");
-      setGetAudioStats(getAudioStats);
-    }
-  }, [setGetAudioStats, getAudioStats]);
-
-  useEffect(() => {
-    decoderWorker.current.onmessage = onWorkerMessage;
-    // 960 = 24000 / 12.5 / 2
-    // The /2 is a bit optional, but won't hurt for recording the mic, and for the
-    // the decoding it might help getting some decoded audio out asap.
-    decoderWorker.current.postMessage({
-      command: "init",
-      bufferLength: 960 * audioContext.current.sampleRate / 24000,
-      decoderSampleRate: 24000,
-      outputBufferSampleRate: audioContext.current.sampleRate,
-      resampleQuality: 0,
-    });
-
-    return () => {
-      console.log("Terminating worker");
-    };
-  }, [onWorkerMessage]);
-
-  return {
-    decodeAudio,
-    analyser,
-    getAudioStats,
-    hasCriticalDelay,
-    setHasCriticalDelay,
-  };
-};
