--- src/features/moshi/hooks/useServerText.ts
@@ -1,32 +1,0 @@
-import { useCallback, useEffect, useState } from "react";
-import { useSocketContext } from "@/features/moshi/hooks/SocketContext";
-import { decodeMessage } from "@/features/moshi/protocol/encoder";
-
-export const useServerText = () => {
-  const [text, setText] = useState<string[]>([]);
-  const [totalTextMessages, setTotalTextMessages] = useState(0);
-  const { socket } = useSocketContext();
-
-  const onSocketMessage = useCallback((e: MessageEvent) => {
-    const dataArray = new Uint8Array(e.data);
-    const message = decodeMessage(dataArray);
-    if (message.type === "text") {
-      setText(text => [...text, message.data]);
-      setTotalTextMessages(count => count + 1);
-    }
-  }, []);
-
-  useEffect(() => {
-    const currentSocket = socket;
-    if (!currentSocket) {
-      return;
-    }
-    setText([]);
-    currentSocket.addEventListener("message", onSocketMessage);
-    return () => {
-      currentSocket.removeEventListener("message", onSocketMessage);
-    };
-  }, [socket]);
-
-  return { text, totalTextMessages };
-};
