--- src/features/moshi2/modules.d.ts
@@ -1,1 +1,0 @@
-declare module "opus-recorder";
